主功能完成

+++ assets/js/admin.js
@@ -6,28 +6,68 @@ const adminLoginBtn=document.querySelector(".adminLoginBtn");
 const _url="https://dealmealserver.onrender.com";
 // const _url="http://localhost:3000";
 
+//sweetalert2
+const Toast = Swal.mixin({
+    toast: true,
+    position: "top",
+    showConfirmButton: false,
+    timer: 1000,
+    timerProgressBar: false,
+    didOpen: (toast) => {
+        toast.onmouseenter = Swal.stopTimer;
+        toast.onmouseleave = Swal.resumeTimer;
+    }
+});
+
+init();
+
+function init(){
+    if( localStorage.getItem("adminToken") ){
+        window.location.href ="admin-products.html";
+    }
+}
+
 function apiAdminLogin(){
     axios.post(`${_url}/login`,{
         "email":adminEmail.value,
         "password":adminPassword.value
     })
     .then(function(res){
-        alert("登入成功");
-        localStorage.setItem("adminToken",res.data.accessToken);
-        localStorage.setItem("adminId",res.data.user.id);
-        window.location.href ="admin-products.html";
+        if(res.data.user.role == "admin"){
+            //sweetalert2
+            Toast.fire({
+                icon: "success",
+                title: "登入成功"
+            }).then((result) => {
+                localStorage.setItem("adminToken",res.data.accessToken);
+                localStorage.setItem("adminId",res.data.user.id);
+                window.location.href ="admin-products.html";
+            });
+            
+        }else{
+            Toast.fire({
+                icon: "error",
+                title: "登入失敗"
+            });
+        }
         
     })
     .catch(function(err){
         console.log(err);
-        alert("登入失敗");
+        Toast.fire({
+            icon: "error",
+            title: "登入失敗"
+        });
     })
 }
 
 adminLoginBtn.addEventListener("click",function(e){
     e.preventDefault();
     if(! (adminEmail.value && adminPassword.value)){
-        alert("欄位未填");
+        Toast.fire({
+            icon: "warning",
+            title: "欄位未填"
+        });
         return;
     }
     apiAdminLogin();
